backendfinanzas/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
numpy==1.26.0
numpy-financial==1.0.0
django-cors-headers==4.3.0
drf-yasg==1.21.7